If movie detail is null, do not let the user bookmark the movie

+++ composeApp/src/commonMain/kotlin/org/ernisernis/ratemoviescmp/movie/presentation/movie_detail/MovieDetailViewModel.kt
@@ -55,6 +55,8 @@ class MovieDetailViewModel(
                 }
             }
             is MovieDetailAction.OnBookmarkClick -> {
+                if (state.value.movie?.movieDetail == null) return
+
                 viewModelScope.launch {
                     if (state.value.isBookmarked) {
                         movieRepository.deleteFromBookmark(movieId)
